JCR-5151 Provide an interface for ConsistencyCheck
* allow plugging in of custom consistency check implementations
* make required methods and classes protected (where possible) or public
* ConsistencyCheck class is deprecated and delegates to DefaultConsistencyCheck
